refactor : Distance 객체로 변환
-리뷰 내용 반영
