+++ src/main/java/nextstep/subway/line/ui/LineController.java
@@ -1,5 +1,6 @@
 package nextstep.subway.line.ui;
 
+import nextstep.subway.exception.InputDataErrorException;
 import nextstep.subway.line.application.LineService;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
@@ -62,6 +63,11 @@ public class LineController {
 
     @ExceptionHandler(DataIntegrityViolationException.class)
     public ResponseEntity handleIllegalArgsException(DataIntegrityViolationException e) {
+        return ResponseEntity.badRequest().build();
+    }
+
+    @ExceptionHandler(InputDataErrorException.class)
+    public ResponseEntity handleIllegalArgsException(InputDataErrorException e) {
         return ResponseEntity.badRequest().build();
     }
 }
